+++ GenericModConfigMenu/Framework/OwnModConfig.cs
@@ -8,10 +8,10 @@ namespace GenericModConfigMenu.Framework
         /*********
         ** Accessors
         *********/
-        /// <summary>The number of field rows to offset when scrolling a config menu.</summary>
-        public int ScrollSpeed { get; set; } = 120;
-
         /// <summary>A keybind which opens the menu.</summary>
         public KeybindList OpenMenuKey = new KeybindList(StardewModdingAPI.SButton.None);
+
+        /// <summary>The number of field rows to offset when scrolling a config menu.</summary>
+        public int ScrollSpeed { get; set; } = 120;
     }
 }